get rid of array and get first element

+++ backend/api/user/profilePic.js
@@ -56,10 +56,12 @@ const handleChangeProfilePictueRequest = async (req, res) => {
 
   await updateProfilePicture(req.user.userID, userFilename);
 
-  try {
-    await fs.rm(path.join(__dirname, `../../profilePics/${oldFile}`));
-  } catch (err) {
-    throw new DatabaseError('Could not remove old profile picture.', 500);
+  if (oldFile != undefined) {
+    try {
+      await fs.rm(path.join(__dirname, `../../profilePics/${oldFile}`));
+    } catch (err) {
+      throw new DatabaseError('Could not remove old profile picture.', 500);
+    }
   }
 
   res.status(200).json({
